requirements.txt
google-generativeai
py-cord
Pybooru